+++ p-login.php
@@ -176,7 +176,7 @@ Password<br>
 <input id=logmein name=logmein class=rb type=submit name=login value=submit><br><br></td>
 <td class=err><?php echo $err;?></td></tr>
 </table>
-<div class=cp>Version 0.9.1<span>&copy;2011 Paul Halliday</span></div>
+<div class=cp>Version 0.9.2<span>&copy;2011 Paul Halliday</span></div>
 </form>
 <script type="text/javascript">document.credcheck.<?php echo $focus;?>.focus();</script>
 </body>